fix(manufacture): restore refactored pipeline after merge conflict
The merge conflict resolution replaced the refactored FlexiManufactureClient
with the old 8-arg inline implementation from main, while the test file
remained on the new version.

Changes:
- ManufactureWorkItems.cs: add ConsumptionResult and ConsolidatedConsumptionCodes records
- IFlexiManufactureDocumentService: add SubmitConsumptionAsync/SubmitProductionAsync (aggregated path) with proper return types
- FlexiManufactureDocumentService: implement all 4 interface methods with doc code capture
- FlexiManufactureClient: restore 9-arg constructor, delegate to injected services
- FlexiManufactureClientTests: replace dead _mockBomClient.Setup with _mockTemplateService.Setup for mixed-ingredient test
