+++ backend/src/Adapters/Anela.Heblo.Adapters.Flexi/Manufacture/Internal/IFlexiManufactureDocumentService.cs
@@ -4,14 +4,26 @@ namespace Anela.Heblo.Adapters.Flexi.Manufacture.Internal;
 
 internal interface IFlexiManufactureDocumentService
 {
-    Task SubmitConsolidatedConsumptionAsync(
+    // Per-product (Product manufacture) path
+    Task<ConsolidatedConsumptionCodes> SubmitConsolidatedConsumptionAsync(
         SubmitManufactureClientRequest request,
         List<ConsumptionItem> consumptionItems,
         Dictionary<string, double> productCosts,
         CancellationToken cancellationToken);
 
-    Task SubmitConsolidatedProductionAsync(
+    Task<string?> SubmitConsolidatedProductionAsync(
         SubmitManufactureClientRequest request,
         Dictionary<string, double> productCosts,
+        CancellationToken cancellationToken);
+
+    // Aggregated (SemiProduct manufacture) path
+    Task<ConsumptionResult> SubmitConsumptionAsync(
+        SubmitManufactureClientRequest request,
+        List<ConsumptionItem> consumptionItems,
+        CancellationToken cancellationToken);
+
+    Task<string?> SubmitProductionAsync(
+        SubmitManufactureClientRequest request,
+        double totalConsumptionCost,
         CancellationToken cancellationToken);
 }
